Logging_Python/requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
docopt==0.6.2
et_xmlfile==2.0.0
fonttools==4.60.1
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.5
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pipreqs==0.4.13
pyparsing==3.2.5
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
yarg==0.1.10
